Move hover impls to lsp module
