+++ rust/kcl-lib/src/parsing/ast/types/mod.rs
@@ -14,7 +14,7 @@ use parse_display::{Display, FromStr};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use tower_lsp::lsp_types::{
-    CompletionItem, CompletionItemKind, DocumentSymbol, FoldingRange, FoldingRangeKind, Range as LspRange, SymbolKind,
+    CompletionItem, CompletionItemKind, DocumentSymbol, FoldingRange, FoldingRangeKind, SymbolKind,
 };
 
 pub use crate::parsing::ast::types::{
@@ -307,23 +307,6 @@ impl Program {
             return false;
         };
         matches!(last, BodyItem::ExpressionStatement(_))
-    }
-
-    pub(crate) fn get_hover_value_for_position(&self, pos: usize, code: &str, scope: Option<Scope>) -> Option<Hover> {
-        // Check if we are in shebang.
-        if let Some(node) = &self.shebang {
-            if node.contains(pos) {
-                let source_range: SourceRange = node.into();
-                return Some(Hover::Comment {
-                    value: r#"The `#!` at the start of a script, known as a shebang, specifies the path to the interpreter that should execute the script. This line is not necessary for your `kcl` to run in the modeling-app. You can safely delete it. If you wish to learn more about what you _can_ do with a shebang, read this doc: [zoo.dev/docs/faq/shebang](https://zoo.dev/docs/faq/shebang)."#.to_string(),
-                    range: source_range.to_lsp_range(code),
-                });
-            }
-        }
-
-        let value = self.get_expr_for_position(pos)?;
-
-        value.get_hover_value_for_position(pos, code, scope)
     }
 
     /// Returns the body item that includes the given character position.
@@ -798,54 +781,6 @@ impl Expr {
         }
     }
 
-    /// Returns a hover value that includes the given character position.
-    /// This is really recursive so keep that in mind.
-    pub(crate) fn get_hover_value_for_position(&self, pos: usize, code: &str, scope: Option<Scope>) -> Option<Hover> {
-        match self {
-            Expr::BinaryExpression(binary_expression) => {
-                binary_expression.get_hover_value_for_position(pos, code, scope)
-            }
-            Expr::FunctionExpression(function_expression) => {
-                function_expression.get_hover_value_for_position(pos, code, scope)
-            }
-            Expr::CallExpression(call_expression) => call_expression.get_hover_value_for_position(pos, code, scope),
-            Expr::CallExpressionKw(call_expression) => call_expression.get_hover_value_for_position(pos, code, scope),
-            Expr::PipeExpression(pipe_expression) => pipe_expression.get_hover_value_for_position(pos, code, scope),
-            Expr::ArrayExpression(array_expression) => array_expression.get_hover_value_for_position(pos, code, scope),
-            Expr::ArrayRangeExpression(array_range) => array_range.get_hover_value_for_position(pos, code, scope),
-            Expr::ObjectExpression(object_expression) => {
-                object_expression.get_hover_value_for_position(pos, code, scope)
-            }
-            Expr::MemberExpression(member_expression) => {
-                member_expression.get_hover_value_for_position(pos, code, scope)
-            }
-            Expr::UnaryExpression(unary_expression) => unary_expression.get_hover_value_for_position(pos, code, scope),
-            Expr::IfExpression(expr) => expr.get_hover_value_for_position(pos, code, scope),
-            // TODO: LSP hover information for values/types. https://github.com/KittyCAD/modeling-app/issues/1126
-            Expr::None(_) => None,
-            Expr::Literal(_) => None,
-            Expr::Identifier(id) => {
-                if id.contains(pos) {
-                    let name = id.name.clone();
-                    Some(Hover::Variable {
-                        ty: scope.and_then(|sc| sc.vars.get(&name).and_then(Clone::clone)),
-                        name,
-                        range: id.as_source_range().to_lsp_range(code),
-                    })
-                } else {
-                    None
-                }
-            }
-            Expr::TagDeclarator(_) => None,
-            // TODO LSP hover info for tag
-            Expr::LabelledExpression(expr) => expr.expr.get_hover_value_for_position(pos, code, scope),
-            // TODO LSP hover info for type
-            Expr::AscribedExpression(expr) => expr.expr.get_hover_value_for_position(pos, code, scope),
-            // TODO: LSP hover information for symbols. https://github.com/KittyCAD/modeling-app/issues/1127
-            Expr::PipeSubstitution(_) => None,
-        }
-    }
-
     /// Rename all identifiers that have the old name to the new given name.
     fn rename_identifiers(&mut self, old_name: &str, new_name: &str) {
         match self {
@@ -1128,30 +1063,6 @@ impl BinaryPart {
             BinaryPart::UnaryExpression(unary_expression) => unary_expression.end,
             BinaryPart::MemberExpression(member_expression) => member_expression.end,
             BinaryPart::IfExpression(e) => e.end,
-        }
-    }
-
-    /// Returns a hover value that includes the given character position.
-    fn get_hover_value_for_position(&self, pos: usize, code: &str, scope: Option<Scope>) -> Option<Hover> {
-        match self {
-            BinaryPart::Literal(_literal) => None,
-            BinaryPart::Identifier(_identifier) => None,
-            BinaryPart::BinaryExpression(binary_expression) => {
-                binary_expression.get_hover_value_for_position(pos, code, scope)
-            }
-            BinaryPart::CallExpression(call_expression) => {
-                call_expression.get_hover_value_for_position(pos, code, scope)
-            }
-            BinaryPart::CallExpressionKw(call_expression) => {
-                call_expression.get_hover_value_for_position(pos, code, scope)
-            }
-            BinaryPart::UnaryExpression(unary_expression) => {
-                unary_expression.get_hover_value_for_position(pos, code, scope)
-            }
-            BinaryPart::IfExpression(e) => e.get_hover_value_for_position(pos, code, scope),
-            BinaryPart::MemberExpression(member_expression) => {
-                member_expression.get_hover_value_for_position(pos, code, scope)
-            }
         }
     }
 
@@ -1730,30 +1641,6 @@ impl CallExpression {
         }
     }
 
-    /// Returns a hover value that includes the given character position.
-    fn get_hover_value_for_position(&self, pos: usize, code: &str, _scope: Option<Scope>) -> Option<Hover> {
-        let callee_source_range: SourceRange = self.callee.clone().into();
-        if callee_source_range.contains(pos) {
-            return Some(Hover::Function {
-                name: self.callee.name.clone(),
-                range: callee_source_range.to_lsp_range(code),
-            });
-        }
-
-        for (index, arg) in self.arguments.iter().enumerate() {
-            let source_range: SourceRange = arg.into();
-            if source_range.contains(pos) {
-                return Some(Hover::Signature {
-                    name: self.callee.name.clone(),
-                    parameter_index: index as u32,
-                    range: source_range.to_lsp_range(code),
-                });
-            }
-        }
-
-        None
-    }
-
     /// Rename all identifiers that have the old name to the new given name.
     fn rename_identifiers(&mut self, old_name: &str, new_name: &str) {
         self.callee.rename(old_name, new_name);
@@ -1791,30 +1678,6 @@ impl CallExpressionKw {
         for arg in &mut self.arguments {
             arg.arg.replace_value(source_range, new_value.clone());
         }
-    }
-
-    /// Returns a hover value that includes the given character position.
-    fn get_hover_value_for_position(&self, pos: usize, code: &str, _scope: Option<Scope>) -> Option<Hover> {
-        let callee_source_range: SourceRange = self.callee.clone().into();
-        if callee_source_range.contains(pos) {
-            return Some(Hover::Function {
-                name: self.callee.name.clone(),
-                range: callee_source_range.to_lsp_range(code),
-            });
-        }
-
-        for (index, arg) in self.iter_arguments().enumerate() {
-            let source_range: SourceRange = arg.into();
-            if source_range.contains(pos) {
-                return Some(Hover::Signature {
-                    name: self.callee.name.clone(),
-                    parameter_index: index as u32,
-                    range: source_range.to_lsp_range(code),
-                });
-            }
-        }
-
-        None
     }
 
     /// Rename all identifiers that have the old name to the new given name.
@@ -2363,18 +2226,6 @@ impl ArrayExpression {
         }
     }
 
-    /// Returns a hover value that includes the given character position.
-    fn get_hover_value_for_position(&self, pos: usize, code: &str, scope: Option<Scope>) -> Option<Hover> {
-        for element in &self.elements {
-            let element_source_range: SourceRange = element.into();
-            if element_source_range.contains(pos) {
-                return element.get_hover_value_for_position(pos, code, scope);
-            }
-        }
-
-        None
-    }
-
     /// Rename all identifiers that have the old name to the new given name.
     fn rename_identifiers(&mut self, old_name: &str, new_name: &str) {
         for element in &mut self.elements {
@@ -2426,18 +2277,6 @@ impl ArrayRangeExpression {
     pub fn replace_value(&mut self, source_range: SourceRange, new_value: Expr) {
         self.start_element.replace_value(source_range, new_value.clone());
         self.end_element.replace_value(source_range, new_value.clone());
-    }
-
-    /// Returns a hover value that includes the given character position.
-    fn get_hover_value_for_position(&self, pos: usize, code: &str, scope: Option<Scope>) -> Option<Hover> {
-        for element in [&self.start_element, &self.end_element] {
-            let element_source_range: SourceRange = element.into();
-            if element_source_range.contains(pos) {
-                return element.get_hover_value_for_position(pos, code, scope);
-            }
-        }
-
-        None
     }
 
     /// Rename all identifiers that have the old name to the new given name.
@@ -2492,18 +2331,6 @@ impl ObjectExpression {
         }
     }
 
-    /// Returns a hover value that includes the given character position.
-    fn get_hover_value_for_position(&self, pos: usize, code: &str, scope: Option<Scope>) -> Option<Hover> {
-        for property in &self.properties {
-            let property_source_range: SourceRange = property.into();
-            if property_source_range.contains(pos) {
-                return property.get_hover_value_for_position(pos, code, scope);
-            }
-        }
-
-        None
-    }
-
     /// Rename all identifiers that have the old name to the new given name.
     fn rename_identifiers(&mut self, old_name: &str, new_name: &str) {
         for property in &mut self.properties {
@@ -2552,16 +2379,6 @@ impl ObjectProperty {
             digest: None,
         })
     }
-
-    /// Returns a hover value that includes the given character position.
-    fn get_hover_value_for_position(&self, pos: usize, code: &str, scope: Option<Scope>) -> Option<Hover> {
-        let value_source_range: SourceRange = self.value.clone().into();
-        if value_source_range.contains(pos) {
-            return self.value.get_hover_value_for_position(pos, code, scope);
-        }
-
-        None
-    }
 }
 
 #[derive(Debug, Clone, Deserialize, Serialize, PartialEq, ts_rs::TS, JsonSchema)]
@@ -2573,16 +2390,6 @@ pub enum MemberObject {
 }
 
 impl MemberObject {
-    /// Returns a hover value that includes the given character position.
-    fn get_hover_value_for_position(&self, pos: usize, code: &str, scope: Option<Scope>) -> Option<Hover> {
-        match self {
-            MemberObject::MemberExpression(member_expression) => {
-                member_expression.get_hover_value_for_position(pos, code, scope)
-            }
-            MemberObject::Identifier(_identifier) => None,
-        }
-    }
-
     pub fn start(&self) -> usize {
         match self {
             MemberObject::MemberExpression(member_expression) => member_expression.start,
@@ -2670,16 +2477,6 @@ impl Node<MemberExpression> {
 }
 
 impl MemberExpression {
-    /// Returns a hover value that includes the given character position.
-    fn get_hover_value_for_position(&self, pos: usize, code: &str, scope: Option<Scope>) -> Option<Hover> {
-        let object_source_range: SourceRange = self.object.clone().into();
-        if object_source_range.contains(pos) {
-            return self.object.get_hover_value_for_position(pos, code, scope);
-        }
-
-        None
-    }
-
     /// Rename all identifiers that have the old name to the new given name.
     fn rename_identifiers(&mut self, old_name: &str, new_name: &str) {
         match &mut self.object {
@@ -2746,22 +2543,6 @@ impl BinaryExpression {
 
     pub fn precedence(&self) -> u8 {
         self.operator.precedence()
-    }
-
-    /// Returns a hover value that includes the given character position.
-    fn get_hover_value_for_position(&self, pos: usize, code: &str, scope: Option<Scope>) -> Option<Hover> {
-        let left_source_range: SourceRange = self.left.clone().into();
-        let right_source_range: SourceRange = self.right.clone().into();
-
-        if left_source_range.contains(pos) {
-            return self.left.get_hover_value_for_position(pos, code, scope);
-        }
-
-        if right_source_range.contains(pos) {
-            return self.right.get_hover_value_for_position(pos, code, scope);
-        }
-
-        None
     }
 
     /// Rename all identifiers that have the old name to the new given name.
@@ -2925,16 +2706,6 @@ impl UnaryExpression {
         self.argument.get_constraint_level()
     }
 
-    /// Returns a hover value that includes the given character position.
-    fn get_hover_value_for_position(&self, pos: usize, code: &str, scope: Option<Scope>) -> Option<Hover> {
-        let argument_source_range: SourceRange = self.argument.clone().into();
-        if argument_source_range.contains(pos) {
-            return self.argument.get_hover_value_for_position(pos, code, scope);
-        }
-
-        None
-    }
-
     /// Rename all identifiers that have the old name to the new given name.
     fn rename_identifiers(&mut self, old_name: &str, new_name: &str) {
         self.argument.rename_identifiers(old_name, new_name);
@@ -3017,18 +2788,6 @@ impl PipeExpression {
         for value in &mut self.body {
             value.replace_value(source_range, new_value.clone());
         }
-    }
-
-    /// Returns a hover value that includes the given character position.
-    fn get_hover_value_for_position(&self, pos: usize, code: &str, scope: Option<Scope>) -> Option<Hover> {
-        for b in &self.body {
-            let b_source_range: SourceRange = b.into();
-            if b_source_range.contains(pos) {
-                return b.get_hover_value_for_position(pos, code, scope);
-            }
-        }
-
-        None
     }
 
     /// Rename all identifiers that have the old name to the new given name.
@@ -3298,20 +3057,6 @@ impl FunctionExpression {
         self.body.replace_value(source_range, new_value);
     }
 
-    /// Returns a hover value that includes the given character position.
-    fn get_hover_value_for_position(&self, pos: usize, code: &str, scope: Option<Scope>) -> Option<Hover> {
-        if let Some(value) = self.body.get_expr_for_position(pos) {
-            let mut scope = scope.unwrap_or_else(|| Scope::default());
-            for arg in &self.params {
-                let ty = arg.type_.as_ref().map(|ty| ty.recast(&FormatOptions::default(), 0));
-                scope.vars.insert(arg.identifier.inner.name.clone(), ty);
-            }
-            return value.get_hover_value_for_position(pos, code, Some(scope));
-        }
-
-        None
-    }
-
     pub fn signature(&self) -> String {
         let mut signature = String::new();
 
@@ -3364,35 +3109,6 @@ pub struct ReturnStatement {
     #[serde(default, skip_serializing_if = "Option::is_none")]
     #[ts(optional)]
     pub digest: Option<Digest>,
-}
-
-/// Describes information about a hover.
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub enum Hover {
-    Function {
-        name: String,
-        range: LspRange,
-    },
-    Signature {
-        name: String,
-        parameter_index: u32,
-        range: LspRange,
-    },
-    Comment {
-        value: String,
-        range: LspRange,
-    },
-    Variable {
-        name: String,
-        ty: Option<String>,
-        range: LspRange,
-    },
-}
-
-#[derive(Debug, Clone, Default)]
-pub(crate) struct Scope {
-    vars: HashMap<String, Option<String>>,
 }
 
 /// Format options.
